+++ ofdb-gateways/src/email/mailgun.rs
@@ -1,24 +1,29 @@
-#[cfg(not(test))]
-use std::io::{Error, ErrorKind};
-use std::{io::Result, thread};
-
+use anyhow::Result;
 use itertools::Itertools;
 use ofdb_core::gateways::email::EmailGateway;
 use ofdb_entities::email::*;
+use std::thread;
 
 /// An email notification manager based on mailgun.net.
 #[derive(Debug, Clone)]
 pub struct Mailgun {
     pub api_key: String,
-    pub api_url: String,
+    pub api_base_url: String,
     pub domain: String,
     pub from_email: EmailAddress,
 }
 
 impl Mailgun {
     fn send(&self, params: Vec<(&'static str, String)>) {
-        let url = self.api_url.clone();
-        let key = self.api_key.clone();
+        let Self {
+            api_base_url,
+            domain,
+            api_key,
+            ..
+        } = &self;
+        // TODO: use url::Url
+        let url = format!("{api_base_url}/{domain}/messages");
+        let key = api_key.clone();
         // TODO: use tokio::task::spawn_blocking
         thread::spawn(move || {
             if let Err(err) = send_raw(&url, &key, params) {
@@ -28,28 +33,27 @@ impl Mailgun {
     }
 }
 
+#[derive(Debug, serde::Deserialize, thiserror::Error)]
+#[error("{message}")]
+struct JsonError {
+    pub message: String,
+}
+
 #[cfg(not(test))]
 fn send_raw(url: &str, api_key: &str, params: Vec<(&'static str, String)>) -> Result<()> {
     let client = reqwest::blocking::Client::new();
-    let res = client
+    let response = client
         .post(url)
         .form(&params)
         .basic_auth("api", Some(api_key))
-        .send();
-    res.map_err(|err| Error::new(ErrorKind::Other, err))
-        .and_then(|res| {
-            if res.status().is_success() {
-                log::debug!("Mail provider response: {:#?}", res);
-                Ok(())
-            } else {
-                // TODO: We should distinguish between a technical failure (Err, see above)
-                // and an error response (here).
-                Err(Error::new(
-                    ErrorKind::Other,
-                    format!("Could not send email: response status: {:?}", res.status()),
-                ))
-            }
-        })
+        .send()?;
+    if response.status().is_success() {
+        log::debug!("Mail provider response: {:#?}", response);
+        Ok(())
+    } else {
+        let json_error: JsonError = response.json()?;
+        Err(json_error.into())
+    }
 }
 
 /// Don't actually send emails while running the tests or
@@ -66,10 +70,15 @@ impl EmailGateway for Mailgun {
             log::warn!("No valid email addresses specified");
             return;
         }
-        log::debug!("Sending e-mails to: {:?}", recipients);
+        log::debug!(
+            "Sending e-mails from {} to: {:?}",
+            self.from_email,
+            recipients
+        );
         let recipients: String = recipients.iter().map(EmailAddress::as_str).join(",");
 
         let params = vec![
+            ("to", self.from_email.as_str().to_owned()), // `to` is required
             ("from", self.from_email.as_str().to_owned()),
             ("bcc", recipients),
             ("subject", email.subject.to_owned()),